organize http code
