+++ dfir_rs/src/util/http/tests/mod.rs
@@ -1,0 +1,1 @@
+pub mod integration_tests;
